ssl-crawler/requirements.txt
python-dotenv==0.19.2
botocore==1.23.17
boto3==1.20.17
alembic==1.7.5
psycopg2-binary==2.9.2
prometheus_client==0.12.0
sqlalchemy==1.4.27
sslyze==5.0.0
python-json-logger==2.0.2
psutil==5.8.0
cryptography~=35.0.0
unittest-xml-reporting==3.0.4

# For tests
docker==5.0.3
